refactor(components) : group navbar components
