+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from 'next';
 import { Roboto } from 'next/font/google';
 import RecoilRootWrapper from '@/recoil/recoil_wrapper';
 import { Suspense } from 'react';
-import Navbar from '../components/Navbar';
+import Navbar from '../components/navbar/Navbar';
 import Loading from '../components/Loading';
 import { Toaster } from 'react-hot-toast';
